backend/pyproject.toml
[build-system]
requires = ["setuptools>=70.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-dev-guardian"
version = "0.1.0"
description = "AI Developer Governance & Codebase Evaluator — A GraphRAG-powered multi-agent system that autonomously evaluates, tests, and governs AI-generated code against proprietary codebases."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Smayan Kulkarni"}
]

dependencies = [
    # CLI Framework
    "typer[all]>=0.12.0",
    # Data Validation & Settings
    "pydantic>=2.6.0",
    "pydantic-settings>=2.0.0",
    # AST Parsing
    "tree-sitter>=0.22.0",
    "tree-sitter-python>=0.23.0",
    # Structured Logging
    "structlog>=24.1.0",
    # Environment Management
    "python-dotenv>=1.0.0",
    # LLM Provider (Groq)
    "groq>=0.9.0",
    # LLMOps Governance
    "langfuse>=2.40.0",
    # Phase 2: Graph & Vector Databases
    "gqlalchemy>=1.6.0",
    "qdrant-client>=1.9.0",
    "fastembed>=0.3.0",
    # Phase 3: Agent Orchestration
    "langgraph>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[project.scripts]
guardian = "dev_guardian.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
pythonPlatform = "All"

[tool.mypy]
python_version = "3.11"
strict = true
